:recycle: Extract issue create form scripts to composable
