+++ issuepoker-frontend/src/types/ComposableParam.ts
@@ -1,0 +1,3 @@
+import type { Ref } from "vue";
+
+export type ComposableParam<T> = T | Ref<T> | (() => T);
